+++ kira/src/arrangement/handle.rs
@@ -1,10 +1,10 @@
 //! An interface for controlling arrangements.
 
-use flume::Sender;
-use thiserror::Error;
-
 use crate::{
-	command::{Command, InstanceCommand},
+	command::{
+		producer::{CommandError, CommandProducer},
+		InstanceCommand,
+	},
 	instance::{
 		handle::InstanceHandle, Instance, InstanceSettings, PauseInstanceSettings,
 		ResumeInstanceSettings, StopInstanceSettings,
@@ -14,15 +14,6 @@ use crate::{
 
 use super::{Arrangement, ArrangementId};
 
-/// Something that can go wrong when using an [`ArrangementHandle`] to
-/// control a arrangement.
-#[derive(Debug, Error)]
-pub enum ArrangementHandleError {
-	/// The audio thread has finished and can no longer receive commands.
-	#[error("The backend cannot receive commands because it no longer exists")]
-	BackendDisconnected,
-}
-
 /// Allows you to control an arrangement.
 #[derive(Debug, Clone)]
 pub struct ArrangementHandle {
@@ -31,11 +22,11 @@ pub struct ArrangementHandle {
 	default_track: TrackIndex,
 	semantic_duration: Option<f64>,
 	default_loop_start: Option<f64>,
-	command_sender: Sender<Command>,
+	command_sender: CommandProducer,
 }
 
 impl ArrangementHandle {
-	pub(crate) fn new(arrangement: &Arrangement, command_sender: Sender<Command>) -> Self {
+	pub(crate) fn new(arrangement: &Arrangement, command_sender: CommandProducer) -> Self {
 		Self {
 			id: arrangement.id(),
 			duration: arrangement.duration(),
@@ -76,10 +67,7 @@ impl ArrangementHandle {
 	}
 
 	/// Plays the arrangement.
-	pub fn play(
-		&mut self,
-		settings: InstanceSettings,
-	) -> Result<InstanceHandle, ArrangementHandleError> {
+	pub fn play(&mut self, settings: InstanceSettings) -> Result<InstanceHandle, CommandError> {
 		let id = settings.id;
 		let instance = Instance::new(
 			self.id.into(),
@@ -89,32 +77,25 @@ impl ArrangementHandle {
 		);
 		let handle = InstanceHandle::new(id, instance.public_state(), self.command_sender.clone());
 		self.command_sender
-			.send(InstanceCommand::Play(id, instance).into())
-			.map_err(|_| ArrangementHandleError::BackendDisconnected)?;
+			.push(InstanceCommand::Play(id, instance).into())?;
 		Ok(handle)
 	}
 
 	/// Pauses all instances of this arrangement.
-	pub fn pause(&mut self, settings: PauseInstanceSettings) -> Result<(), ArrangementHandleError> {
+	pub fn pause(&mut self, settings: PauseInstanceSettings) -> Result<(), CommandError> {
 		self.command_sender
-			.send(InstanceCommand::PauseInstancesOf(self.id.into(), settings).into())
-			.map_err(|_| ArrangementHandleError::BackendDisconnected)
+			.push(InstanceCommand::PauseInstancesOf(self.id.into(), settings).into())
 	}
 
 	/// Resumes all instances of this arrangement.
-	pub fn resume(
-		&mut self,
-		settings: ResumeInstanceSettings,
-	) -> Result<(), ArrangementHandleError> {
+	pub fn resume(&mut self, settings: ResumeInstanceSettings) -> Result<(), CommandError> {
 		self.command_sender
-			.send(InstanceCommand::ResumeInstancesOf(self.id.into(), settings).into())
-			.map_err(|_| ArrangementHandleError::BackendDisconnected)
+			.push(InstanceCommand::ResumeInstancesOf(self.id.into(), settings).into())
 	}
 
 	/// Stops all instances of this arrangement.
-	pub fn stop(&mut self, settings: StopInstanceSettings) -> Result<(), ArrangementHandleError> {
+	pub fn stop(&mut self, settings: StopInstanceSettings) -> Result<(), CommandError> {
 		self.command_sender
-			.send(InstanceCommand::StopInstancesOf(self.id.into(), settings).into())
-			.map_err(|_| ArrangementHandleError::BackendDisconnected)
+			.push(InstanceCommand::StopInstancesOf(self.id.into(), settings).into())
 	}
 }