+++ kira/src/metronome/handle.rs
@@ -1,42 +1,49 @@
 //! An interface for controlling metronomes.
 
-use flume::{Receiver, Sender, TryIter};
+use std::sync::{Arc, Mutex};
+
+use ringbuf::Consumer;
 use thiserror::Error;
 
 use crate::{
-	command::{Command, MetronomeCommand},
+	command::{
+		producer::{CommandError, CommandProducer},
+		MetronomeCommand,
+	},
 	Tempo, Value,
 };
 
 use super::MetronomeId;
 
 /// Something that can go wrong when using a [`MetronomeHandle`]
-/// to control a metronome.
+/// to receive an event from a metronome.
 #[derive(Debug, Error)]
-pub enum MetronomeHandleError {
-	/// The audio thread has finished and can no longer receive commands.
-	#[error("The backend cannot receive commands because it no longer exists")]
-	BackendDisconnected,
+pub enum PopMetronomeEventError {
+	/// A thread panicked while using the event receiver.
+	#[error("The event receiver cannot be used because a thread panicked while borrowing it.")]
+	MutexPoisoned,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Clone)]
 /// Allows you to control a metronome.
+
+// TODO: add a manual impl of Debug
 pub struct MetronomeHandle {
 	id: MetronomeId,
-	command_sender: Sender<Command>,
-	event_receiver: Receiver<f64>,
+	command_sender: CommandProducer,
+	event_receiver: Arc<Mutex<Consumer<f64>>>,
 }
 
 impl MetronomeHandle {
 	pub(crate) fn new(
 		id: MetronomeId,
-		command_sender: Sender<Command>,
-		event_receiver: Receiver<f64>,
+		command_sender: CommandProducer,
+		event_receiver: Consumer<f64>,
 	) -> Self {
 		Self {
 			id,
 			command_sender,
-			event_receiver,
+			event_receiver: Arc::new(Mutex::new(event_receiver)),
 		}
 	}
 
@@ -46,39 +53,36 @@ impl MetronomeHandle {
 	}
 
 	/// Sets the tempo of the metronome (in beats per minute).
-	pub fn set_tempo(
-		&mut self,
-		tempo: impl Into<Value<Tempo>>,
-	) -> Result<(), MetronomeHandleError> {
+	pub fn set_tempo(&mut self, tempo: impl Into<Value<Tempo>>) -> Result<(), CommandError> {
 		self.command_sender
-			.send(MetronomeCommand::SetMetronomeTempo(self.id(), tempo.into()).into())
-			.map_err(|_| MetronomeHandleError::BackendDisconnected)
+			.push(MetronomeCommand::SetMetronomeTempo(self.id(), tempo.into()).into())
 	}
 
 	/// Starts the metronome.
-	pub fn start(&mut self) -> Result<(), MetronomeHandleError> {
+	pub fn start(&mut self) -> Result<(), CommandError> {
 		self.command_sender
-			.send(MetronomeCommand::StartMetronome(self.id()).into())
-			.map_err(|_| MetronomeHandleError::BackendDisconnected)
+			.push(MetronomeCommand::StartMetronome(self.id()).into())
 	}
 
 	/// Pauses the metronome.
-	pub fn pause(&mut self) -> Result<(), MetronomeHandleError> {
+	pub fn pause(&mut self) -> Result<(), CommandError> {
 		self.command_sender
-			.send(MetronomeCommand::PauseMetronome(self.id()).into())
-			.map_err(|_| MetronomeHandleError::BackendDisconnected)
+			.push(MetronomeCommand::PauseMetronome(self.id()).into())
 	}
 
 	/// Stops the metronome and resets its time to zero.
-	pub fn stop(&mut self) -> Result<(), MetronomeHandleError> {
+	pub fn stop(&mut self) -> Result<(), CommandError> {
 		self.command_sender
-			.send(MetronomeCommand::StopMetronome(self.id()).into())
-			.map_err(|_| MetronomeHandleError::BackendDisconnected)
+			.push(MetronomeCommand::StopMetronome(self.id()).into())
 	}
 
-	/// Returns an iterator over new interval events
-	/// that this metronome emitted.
-	pub fn event_iter(&mut self) -> TryIter<f64> {
-		self.event_receiver.try_iter()
+	/// Gets the first interval event that was emitted by this
+	/// metronome since the last call to `pop_event`.
+	pub fn pop_event(&mut self) -> Result<Option<f64>, PopMetronomeEventError> {
+		Ok(self
+			.event_receiver
+			.lock()
+			.map_err(|_| PopMetronomeEventError::MutexPoisoned)?
+			.pop())
 	}
 }