+++ kira/src/mixer/track/handle.rs
@@ -1,21 +1,23 @@
 //! An interface for controlling mixer tracks.
 
 use basedrop::Owned;
-use flume::Sender;
 use indexmap::IndexSet;
 use thiserror::Error;
 
 use crate::{
-	command::{Command, MixerCommand},
+	command::{
+		producer::{CommandError, CommandProducer},
+		MixerCommand,
+	},
 	mixer::effect::{handle::EffectHandle, Effect, EffectId, EffectSettings},
 };
 
 use super::{TrackIndex, TrackSettings};
 
 /// Something that can go wrong when using a [`TrackHandle`] to
-/// control a mixer track.
+/// add an effect to a mixer track.
 #[derive(Debug, Error)]
-pub enum TrackHandleError {
+pub enum AddEffectError {
 	/// The maximum effect limit for this track has been reached.
 	#[error(
 		"Cannot add an effect because the max number of effects for this track has been reached"
@@ -24,15 +26,27 @@ pub enum TrackHandleError {
 	/// No effect with the specified ID exists on this track.
 	#[error("No effect with the specified ID exists on this track")]
 	NoEffectWithId(EffectId),
-	/// The audio thread has finished and can no longer receive commands.
-	#[error("The backend cannot receive commands because it no longer exists")]
-	BackendDisconnected,
+	/// A command could not be sent to the audio thread.
+	#[error("Could not send the command to the audio thread.")]
+	CommandProducerError(#[from] CommandError),
+}
+
+/// Something that can go wrong when using a [`TrackHandle`] to
+/// remove an effect from a mixer track.
+#[derive(Debug, Error)]
+pub enum RemoveEffectError {
+	/// No effect with the specified ID exists on this track.
+	#[error("No effect with the specified ID exists on this track")]
+	NoEffectWithId(EffectId),
+	/// A command could not be sent to the audio thread.
+	#[error("Could not send the command to the audio thread.")]
+	CommandProducerError(#[from] CommandError),
 }
 
 /// Allows you to control a mixer sound.
 pub struct TrackHandle {
 	index: TrackIndex,
-	command_sender: Sender<Command>,
+	command_sender: CommandProducer,
 	active_effect_ids: IndexSet<EffectId>,
 	resource_collector_handle: basedrop::Handle,
 }
@@ -41,7 +55,7 @@ impl TrackHandle {
 	pub(crate) fn new(
 		index: TrackIndex,
 		settings: &TrackSettings,
-		command_sender: Sender<Command>,
+		command_sender: CommandProducer,
 		resource_collector_handle: basedrop::Handle,
 	) -> Self {
 		Self {
@@ -62,34 +76,32 @@ impl TrackHandle {
 		&mut self,
 		effect: impl Effect + 'static,
 		settings: EffectSettings,
-	) -> Result<EffectHandle, TrackHandleError> {
+	) -> Result<EffectHandle, AddEffectError> {
 		if self.active_effect_ids.len() >= self.active_effect_ids.capacity() {
-			return Err(TrackHandleError::EffectLimitReached);
+			return Err(AddEffectError::EffectLimitReached);
 		}
 		let id = settings.id;
 		let handle = EffectHandle::new(self.index, &settings, self.command_sender.clone());
-		self.command_sender
-			.send(
-				MixerCommand::AddEffect(
-					self.index,
-					Owned::new(&self.resource_collector_handle, Box::new(effect)),
-					settings,
-				)
-				.into(),
+		self.command_sender.push(
+			MixerCommand::AddEffect(
+				self.index,
+				Owned::new(&self.resource_collector_handle, Box::new(effect)),
+				settings,
 			)
-			.map_err(|_| TrackHandleError::BackendDisconnected)?;
+			.into(),
+		)?;
 		self.active_effect_ids.insert(id);
 		Ok(handle)
 	}
 
 	/// Removes an effect from the track.
-	pub fn remove_effect(&mut self, id: impl Into<EffectId>) -> Result<(), TrackHandleError> {
+	pub fn remove_effect(&mut self, id: impl Into<EffectId>) -> Result<(), RemoveEffectError> {
 		let id = id.into();
 		if !self.active_effect_ids.remove(&id) {
-			return Err(TrackHandleError::NoEffectWithId(id));
+			return Err(RemoveEffectError::NoEffectWithId(id));
 		}
 		self.command_sender
-			.send(MixerCommand::RemoveEffect(self.index, id).into())
-			.map_err(|_| TrackHandleError::BackendDisconnected)
+			.push(MixerCommand::RemoveEffect(self.index, id).into())?;
+		Ok(())
 	}
 }